Add `MQTTRoomBase` class
